move redis integration to separate directory
